multi_stage: write commands.sh to configmap
Instead of running commands directly with `bash` when running multistage
tests, `ci-operator` should instead execute the script directly so that
the shebang specifying what shell to use is followed (it is ignored when
run directly by the shell) and to remove the multistage test steps'
dependance on `bash`, which does not exist in certain images, usually
alpine based.

This PR now runs commands by creating a configmap and mounting it to a
known location in the pod. Then, the script is executed by doing
`/bin/sh -c /path/to/script.sh`. At a bare minimum, all images should
have a `/bin/sh`, and in order to continue supporting injecting the cli
binary into images, the ci-operator needs to execute commands via
`/bin/sh` so the `$PATH` env var can be updated to include the path to
the cli before running the user's script.